refactor: move shell tokenizer in a extra file
